requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
DataProperty==1.0.1
datasets==2.18.0
deepspeed==0.14.0
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docstring_parser==0.16
editdistance==0.8.1
email_validator==2.2.0
evaluate==0.4.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.3
fire==0.6.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.37.2
gradio_client==1.0.2
grpcio==1.64.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
loguru==0.7.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mbstrdecoder==1.1.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
orjson==3.10.6
packaging==23.2
pandas==2.2.1
paretoset==1.2.3
pathvalidate==3.2.0
peft==0.12.0
pillow==10.4.0
platformdirs==4.2.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pybind11==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.1
ruff==0.5.1
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.37.2
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.3.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
tqdm==4.66.2
tqdm-multiprocess==0.0.11
transformers==4.44.2
triton==2.1.0
trl==0.9.4
typepy==1.3.2
typer==0.12.3
typing_extensions==4.10.0
tyro==0.8.5
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
word2number==1.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
